requirements.txt
requests
python-dotenv
langchain
openai
